[#90] [don't delete RND file
Former-commit-id: 7b364754dda3c7a7de36870e4e28478364e21f1a

+++ files/tasks.py
@@ -95,9 +95,10 @@ def worker(filename, project, description, vendor, file_type, network, file_id):
     available_percent = percent_per_file / 2
     tables = set()    
     for f in work_files:
-        for f in Files.objects.filter(project=project, file_type=file_type):
-            f.clear_tables()           
-            f.delete()
+        if file_type not in ['RND', ]:            
+            for f in Files.objects.filter(project=project, file_type=file_type):
+                f.clear_tables()           
+                f.delete()
         data_file = None
         if (file_type in xml_types) and (network == 'WCDMA'):
             data_file = WcdmaXML(f, project, file_id, i, available_percent, set_percent).data
